Conflicts:
	src/main/java/fr/insa/binder/projets5/mavenproject1/gui/technicien/technicienInterface/Supp_etat.java
	src/main/java/fr/insa/binder/projets5/mavenproject1/gui/technicien/technicienMachine/Etat_d_une_machine.java
	src/main/java/fr/insa/binder/projets5/mavenproject1/gui/technicien/technicienMachine/Grid_etat_d_une_machine.java

+++ src/main/java/fr/insa/binder/projets5/mavenproject1/type_operation.java
@@ -43,10 +43,12 @@ public class type_operation {
     }
 
     public static void supTypeOperation(Connection con, int id) throws SQLException {
-        try (PreparedStatement pst = con.prepareStatement(
-                "delete from type_operation_bof where id_type_operation = ?")) {
-            pst.setInt(1, id);
-            pst.executeUpdate();
+        if (id!=1){
+            try (PreparedStatement pst = con.prepareStatement(
+                    "delete from type_operation_bof where id_type_operation = ?")) {
+                pst.setInt(1, id);
+                pst.executeUpdate();
+            }
         }
     }
